% Auto-generated: do not edit by hand
\name{myauoMyDashMicRecorderComponent}

\alias{myauoMyDashMicRecorderComponent}

\title{MyDashMicRecorderComponent component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
myauoMyDashMicRecorderComponent(id=NULL, audio=NULL, backgroundColor=NULL, record=NULL,
recorderParams=NULL, strokeColor=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{audio}{Named list. The value displayed in the input.}

\item{backgroundColor}{Character. The value displayed in the input.}

\item{record}{Logical. The value displayed in the input.}

\item{recorderParams}{Named list. The value displayed in the input.}

\item{strokeColor}{Character. The value displayed in the input.}

\item{value}{Named list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

